Restructure move ApplicationConfigurator to cli package and CommonFeatureConfig to tools/common
