+++ src/main/groovy/com/cloudogu/gitops/cli/ApplicationConfigurator.groovy
@@ -1,5 +1,6 @@
-package com.cloudogu.gitops.config
+package com.cloudogu.gitops.cli
 
+import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.utils.FileSystemUtils
 
 import groovy.util.logging.Slf4j
